requirements.txt
streamlit>=1.35.0
streamlit-aggrid>=0.3.4
requests>=2.28.0
requests-cache>=1.0.0
biopython>=1.79
pandas>=1.3.0
openpyxl>=3.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
pypdf>=4.0.0
